requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
python-dotenv==1.0.0
requests==2.31.0
deep-translator==1.11.4
reportlab==4.0.4
gunicorn==21.2.0
